+++ tests/test_build_linkcheck.py
@@ -11,9 +11,10 @@
 import http.server
 import json
 import re
-import threading
 from unittest import mock
 import pytest
+
+from utils import http_server
 
 
 @pytest.mark.sphinx('linkcheck', testroot='linkcheck', freshenv=True)
@@ -110,12 +111,12 @@ def test_anchors_ignored(app, status, warning):
 
 @pytest.mark.sphinx('linkcheck', testroot='linkcheck-localserver', freshenv=True)
 def test_raises_for_invalid_status(app, status, warning):
-    server_thread = HttpServerThread(InternalServerErrorHandler, daemon=True)
-    server_thread.start()
-    try:
+    class InternalServerErrorHandler(http.server.BaseHTTPRequestHandler):
+        def do_GET(self):
+            self.send_error(500, "Internal Server Error")
+
+    with http_server(InternalServerErrorHandler):
         app.builder.build_all()
-    finally:
-        server_thread.terminate()
     content = (app.outdir / 'output.txt').read_text()
     assert content == (
         "index.rst:1: [broken] http://localhost:7777/#anchor: "
@@ -177,22 +178,3 @@ def test_linkcheck_request_headers(app, status, warning):
                 assert headers["X-Secret"] == "open sesami"
             else:
                 assert headers["Accept"] == "text/html,application/xhtml+xml;q=0.9,*/*;q=0.8"
-
-
-class HttpServerThread(threading.Thread):
-    def __init__(self, handler, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.server = http.server.HTTPServer(("localhost", 7777), handler)
-
-    def run(self):
-        self.server.serve_forever(poll_interval=0.01)
-
-    def terminate(self):
-        self.server.shutdown()
-        self.server.server_close()
-        self.join()
-
-
-class InternalServerErrorHandler(http.server.BaseHTTPRequestHandler):
-    def do_GET(self):
-        self.send_error(500, "Internal Server Error")
